rename client command
